New Style Walkers
Walkers behavior is not defined in the class definition.
Processing an AND node always calls walk_and. This makes it possible to
subclass and override methods, but at the same time call the
implementation of a super class, e.g.

  def walk_and(...):
      return ParentWalker.walk_and(self, ....)

The utility method Walker._super is provided to automatically handle the
dispatch of a node to the correct walk_* function, e.g.,:

  def walk_bool_to_bool(...):
      return ParentWalker._super(self, ....)

The method Walker.set_functions is deprecated, but kept for
compatibility with old-style walkers. Using set_functions has the same
effect as before. However, you cannot modify a subclass of a walker
using set_functions. This is ok for the transition phase.

The method Walker.set_handler is used to perform the same operation of
set_function at the class level. The associated decorator @handles can
be used to associate methods with nodetypes.

Other changes:

* ALL_TYPES is now a tuple (immutable)
* new_node_type creates the str representation of the node
* Walker.__init__ iterates over all_types (including custom types), and
  defaults the implementations of "new" nodes to walk_error.

These makes it possible to extend the walkers from outside pySMT,
without relying on hacks like the DWF. An example of this is given in
examples/ltl.py